requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.3
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
PyJWT==2.8.0
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.2
